Mobile improvement for team schedule
